+++ src/common/models/cluster/cluster.mocha.ts
@@ -8,6 +8,10 @@ describe('Cluster', () => {
     testImmutableClass(Cluster, [
       {
         name: 'my-druid-cluster',
+        type: 'druid'
+      },
+      {
+        name: 'my-druid-cluster',
         type: 'druid',
         host: '192.168.99.100',
         version: '0.9.1',
@@ -32,6 +36,22 @@ describe('Cluster', () => {
         database: 'datazoo',
         user: 'datazoo-user',
         password: 'koalas'
+      },
+      {
+        name: 'my-mysql-cluster',
+        type: 'druid',
+        host: '192.168.99.100',
+        timeout: 30000,
+        sourceListScan: 'auto'
+      },
+      {
+        name: 'my-mysql-cluster',
+        type: 'druid',
+        host: '192.168.99.100',
+        timeout: 30000,
+        sourceListScan: 'auto',
+        sourceListRefreshInterval: 0,
+        sourceReintrospectInterval: 0
       }
     ]);
   });