+++ src/server/config.ts
@@ -134,7 +134,13 @@ if (numSettingsInputs === 0) {
 }
 
 if (numSettingsInputs > 1) {
-  exitWithError(`only one of --${SETTINGS_INPUTS.join(', --')} can be given on the command line`);
+  console.error(`only one of --${SETTINGS_INPUTS.join(', --')} can be given on the command line`);
+  if (parsedArgs['druid'] && parsedArgs['config']) {
+    console.error(`Looks like you are using --config and --druid in conjunction with each other`);
+    console.error(`This usage is no longer supported. If you are migrating from Pivot < 0.9.x`);
+    console.error(`Please visit: (https://github.com/implydata/pivot/blob/master/docs/pivot-0.9.x-migration.md)`);
+  }
+  process.exit(1);
 }
 
 export const PRINT_CONFIG = Boolean(parsedArgs['print-config']);