+++ src/server/utils/cluster-manager/cluster-manager.ts
@@ -8,13 +8,14 @@ import { Logger } from '../logger/logger';
 
 const DRUID_REQUEST_DECORATOR_MODULE_VERSION = 1;
 
-export interface RequestDecoratorFactoryOptions {
+export interface RequestDecoratorFactoryParams {
+  options: any;
   cluster: Cluster;
 }
 
 export interface DruidRequestDecoratorModule {
   version: number;
-  druidRequestDecoratorFactory: (logger: Logger, options: RequestDecoratorFactoryOptions) => DruidRequestDecorator;
+  druidRequestDecoratorFactory: (logger: Logger, params: RequestDecoratorFactoryParams) => DruidRequestDecorator;
 }
 
 // For each external we want to maintain its source and weather it should introspect at all
@@ -121,6 +122,7 @@ export class ClusterManager {
     if (cluster.type === 'druid' && requestDecoratorModule) {
       logger.log(`Cluster '${cluster.name}' creating requestDecorator`);
       druidRequestDecorator = requestDecoratorModule.druidRequestDecoratorFactory(logger, {
+        options: cluster.decoratorOptions,
         cluster
       });
     }
@@ -128,7 +130,7 @@ export class ClusterManager {
     this.requester = properRequesterFactory({
       type: cluster.type,
       host: cluster.host,
-      timeout: cluster.timeout,
+      timeout: cluster.getTimeout(),
       verbose: this.verbose,
       concurrentLimit: 5,
 
@@ -143,8 +145,8 @@ export class ClusterManager {
   private updateSourceListRefreshTimer() {
     const { logger, cluster } = this;
 
-    if (this.sourceListRefreshInterval !== cluster.sourceListRefreshInterval) {
-      this.sourceListRefreshInterval = cluster.sourceListRefreshInterval;
+    if (this.sourceListRefreshInterval !== cluster.getSourceListRefreshInterval()) {
+      this.sourceListRefreshInterval = cluster.getSourceListRefreshInterval();
 
       if (this.sourceListRefreshTimer) {
         logger.log(`Clearing sourceListRefresh timer in cluster '${cluster.name}'`);
@@ -165,8 +167,8 @@ export class ClusterManager {
   private updateSourceReintrospectTimer() {
     const { logger, cluster } = this;
 
-    if (this.sourceReintrospectInterval !== cluster.sourceReintrospectInterval) {
-      this.sourceReintrospectInterval = cluster.sourceReintrospectInterval;
+    if (this.sourceReintrospectInterval !== cluster.getSourceReintrospectInterval()) {
+      this.sourceReintrospectInterval = cluster.getSourceReintrospectInterval();
 
       if (this.sourceReintrospectTimer) {
         logger.log(`Clearing sourceReintrospect timer in cluster '${cluster.name}'`);